@@ -1,34 +1,35 @@
 source "https://rubygems.org"
 
-ruby "2.1.2"
+ruby "2.2.0"
 
-gem "active_model_serializers"
+gem "active_model_serializers", "0.8.3"
 gem "analytics-ruby", "~> 2.0.0", require: "segment/analytics"
 gem "angularjs-rails"
+gem "angular_rails_csrf"
 gem "attr_extras"
 gem "bourbon"
 gem "coffee-rails"
 gem "coffeelint"
-gem "dalli"
-gem "faraday-http-cache"
 gem "font-awesome-rails"
 gem "foreman"
 gem "haml-rails"
 gem "high_voltage"
 gem "jquery-rails"
 gem "jshintrb"
-gem "memcachier"
 gem "neat"
 gem "newrelic_rpm"
 gem "octokit"
 gem "omniauth-github"
 gem "paranoia", "~> 2.0"
 gem "pg"
-gem "rails", "4.1.5"
-gem "resque", "~> 1.22.0"
+gem "rails", "4.2"
+gem "responders", "~> 2.0"
+gem "resque", "~> 1.25.0"
 gem "resque-retry"
 gem "rubocop", "0.25.0"
-gem "sass-rails", "~> 4.0.2"
+gem "sass-rails"
+gem "scss-lint", require: false
+gem "sentry-raven"
 gem "stripe"
 gem "uglifier", ">= 1.0.3"
 gem "unicorn"
@@ -39,6 +40,7 @@ end
 
 group :development, :test do
   gem "byebug"
+  gem "jasmine-rails"
   gem "konacha"
   gem "poltergeist"
   gem "rspec-rails", ">= 2.14"
