@@ -1,6 +1,11 @@
 class AccountsController < ApplicationController
   def show
-    @repos = current_user.subscribed_repos.order(:full_github_name)
-    @org_repos, @personal_repos = @repos.partition(&:in_organization?)
+    @account_page = AccountPage.new(find_subscribed_repos)
+  end
+
+  private
+
+  def find_subscribed_repos
+    current_user.subscribed_repos.order(:full_github_name)
   end
 end
