@@ -1,9 +1,10 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery
+
   before_action :force_https
   before_action :capture_campaign_params
   before_action :authenticate
-  after_action  :set_csrf_cookie_for_ng
+
   helper_method :current_user, :signed_in?
 
   private
@@ -30,7 +31,7 @@ class ApplicationController < ActionController::Base
 
   def authenticate
     unless signed_in?
-      redirect_to sign_in_path
+      redirect_to root_path
     end
   end
 
@@ -44,16 +45,6 @@ class ApplicationController < ActionController::Base
 
   def analytics
     @analytics ||= Analytics.new(current_user, session[:campaign_params])
-  end
-
-  def set_csrf_cookie_for_ng
-    if protect_against_forgery?
-      cookies['XSRF-TOKEN'] = form_authenticity_token
-    end
-  end
-
-  def report_exception(exception, metadata)
-    Raven.capture_exception(exception, extra: metadata)
   end
 
   protected
