@@ -6,47 +6,39 @@ class SubscriptionsController < ApplicationController
   respond_to :json
 
   def create
-    if activator.activate(repo, github_token) && create_subscription
+    if activator.activate && create_subscription
       analytics.track_subscribed(repo)
+
       render json: repo, status: :created
     else
-      activator.deactivate(repo, github_token)
-      report_activation_error("Failed to subscribe and activate repo")
+      activator.deactivate
+
       head 502
     end
   end
 
   def destroy
-    repo = current_user.repos.find(params[:repo_id])
-
-    if activator.deactivate(repo, session[:github_token]) && delete_subscription
+    if activator.deactivate && delete_subscription
       analytics.track_unsubscribed(repo)
+
       render json: repo, status: :created
     else
-      report_activation_error("Failed to unsubscribe and deactivate repo")
       head 502
     end
   end
 
   private
 
-  def report_activation_error(message)
-    report_exception(
-      FailedToActivate.new(message),
-      user_id: current_user.id, repo_id: params[:repo_id]
-    )
+  def activator
+    RepoActivator.new(repo: repo, github_token: github_token)
   end
 
   def repo
-    @repo ||= current_user.repos.find(params[:repo_id])
+    @repo ||= current_user.repos.find(params.fetch(:repo_id))
   end
 
   def github_token
-    session[:github_token]
-  end
-
-  def activator
-    RepoActivator.new
+    session.fetch(:github_token)
   end
 
   def create_subscription
