@@ -3,11 +3,11 @@ class RepoInformationJob
 
   @queue = :low
 
-  def self.perform(repo_id, github_token)
+  def self.perform(repo_id)
     repo = Repo.find(repo_id)
     repo.touch
 
-    github = GithubApi.new(github_token)
+    github = GithubApi.new
     github_data = github.repo(repo.full_github_name)
 
     repo.update_attributes!(
@@ -15,6 +15,6 @@ class RepoInformationJob
       in_organization: github_data[:organization].present?
     )
   rescue Resque::TermException
-    Resque.enqueue(self, user_id, github_token)
+    Resque.enqueue(self, repo_id)
   end
 end
