@@ -5,5 +5,5 @@ module Retryable
 
   @retry_limit = 10
   @retry_delay = 120
-  @fatal_exceptions = [Octokit::NotFound, Octokit::Unauthorized]
+  @fatal_exceptions = [Octokit::Unauthorized]
 end
