@@ -5,11 +5,17 @@ class Commit
   def file_content(filename)
     contents = @github.file_contents(repo_name, filename, sha)
     if contents && contents.content
-      Base64.decode64(contents.content)
+      Base64.decode64(contents.content).force_encoding("UTF-8")
     else
       ""
     end
   rescue Octokit::NotFound
     ""
+  rescue Octokit::Forbidden => exception
+    if exception.errors.first[:code] == "too_large"
+      ""
+    else
+      raise exception
+    end
   end
 end
