@@ -12,14 +12,14 @@ class Payload
   end
 
   def github_repo_id
-    data['repository']['id']
+    repository["id"]
   end
 
   def full_repo_name
-    data['repository']['full_name']
+    repository["full_name"]
   end
 
-  def number
+  def pull_request_number
     data['number']
   end
 
@@ -35,6 +35,10 @@ class Payload
     data['zen']
   end
 
+  def repository_owner
+    repository["owner"]["login"]
+  end
+
   private
 
   def parse_data
@@ -47,5 +51,9 @@ class Payload
 
   def pull_request
     data.fetch("pull_request", {})
+  end
+
+  def repository
+    @repository ||= data["repository"]
   end
 end
