@@ -1,29 +1,27 @@
 # Hold file, line, and violation message values.
 # Built by style guides.
 # Printed by Commenter.
-class Violation
-  attr_reader :line_number, :filename
+class Violation < ActiveRecord::Base
+  belongs_to :build
 
-  def initialize(file, line_number, message)
-    @filename = file.filename
-    @line = file.line_at(line_number)
-    @line_number = line_number
-    @messages = [message]
-  end
+  validates :build, presence: true
+  validates :filename, presence: true
 
-  def add_messages(messages)
-    @messages += messages
+  attr_writer :line
+
+  def add_messages(new_messages)
+    self[:messages].concat(new_messages)
   end
 
   def messages
-    @messages.uniq
-  end
-
-  def patch_position
-    @line.patch_position
+    self[:messages].uniq
   end
 
   def on_changed_line?
-    @line.changed?
+    line.changed?
   end
+
+  private
+
+  attr_reader :line
 end
