@@ -20,10 +20,11 @@ class RepoSynchronization
   private
 
   def repo_attributes(attributes)
-    attributes.slice(:private).merge(
+    {
+      private: attributes[:private],
       github_id: attributes[:id],
       full_github_name: attributes[:full_name],
       in_organization: attributes[:owner][:type] == ORGANIZATION_TYPE
-    )
+    }
   end
 end
