@@ -10,6 +10,10 @@ module Houndapp
     config.encoding = "utf-8"
     config.filter_parameters += [:password]
     config.active_support.escape_html_entities_in_json = true
-    config.assets.version = '1.0'
+    config.middleware.insert_before "Rack::ETag", "Rack::Deflater"
+    config.middleware.insert_before(
+      "Rack::ETag",
+      "RedirectToConfiguration"
+    )
   end
 end
