@@ -11,7 +11,7 @@ Houndapp::Application.configure do
   config.action_controller.perform_caching = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = false
+  config.serve_static_files = false
 
   # Compress JavaScripts and CSS
   config.assets.compress = true
@@ -42,21 +42,19 @@ Houndapp::Application.configure do
   # config.logger = ActiveSupport::TaggedLogging.new(SyslogLogger.new)
 
   # Use a different cache store in production
-  config.cache_store = :dalli_store,
-    (ENV["MEMCACHIER_SERVERS"] || "").split(","),
-    {
-      :username => ENV["MEMCACHIER_USERNAME"],
-      :password => ENV["MEMCACHIER_PASSWORD"],
-      :failover => true,
-      :socket_timeout => 1.5,
-      :socket_failure_delay => 0.2
-    }
+  # config.cache_store = :dalli_store,
+  #   (ENV["MEMCACHIER_SERVERS"] || "").split(","),
+  #   {
+  #     :username => ENV["MEMCACHIER_USERNAME"],
+  #     :password => ENV["MEMCACHIER_PASSWORD"],
+  #     :failover => true,
+  #     :socket_timeout => 1.5,
+  #     :socket_failure_delay => 0.2
+  #   }
 
   # Enable serving of images, stylesheets, and JavaScripts from an asset server
   # config.action_controller.asset_host = "http://assets.example.com"
 
-  # Precompile additional assets (application.js, application.css, and all non-JS/CSS are already added)
-  config.assets.precompile += %w( repos/*.js )
   config.assets.js_compressor = :uglifier
 
   # Disable delivery errors, bad email addresses will be ignored
