@@ -1,5 +1,5 @@
 worker_processes Integer(ENV["WEB_CONCURRENCY"] || 3)
-timeout 15
+timeout (ENV["UNICORN_TIMEOUT"] || 15).to_i
 preload_app true
 
 before_fork do |server, worker|
