@@ -1,34 +1,28 @@
-require 'octokit'
-require 'base64'
+require "octokit"
+require "base64"
+require "active_support/core_ext/object/with_options"
 
 class GithubApi
-  SERVICES_TEAM_NAME = 'Services'
+  SERVICES_TEAM_NAME = "Services"
+  PREVIEW_MEDIA_TYPE = "application/vnd.github.moondragon-preview+json"
 
-  pattr_initialize :token
+  def initialize(token = ENV["HOUND_GITHUB_TOKEN"])
+    @token = token
+  end
 
   def client
-    @client ||= Octokit::Client.new(access_token: token)
+    @client ||= Octokit::Client.new(access_token: @token, auto_paginate: true)
   end
 
   def repos
     user_repos + org_repos
   end
 
-  def add_user_to_repo(username, repo_name)
-    repo = repo(repo_name)
-
-    if repo.organization
-      add_user_to_org(username, repo)
-    else
-      client.add_collaborator(repo.full_name, username)
-    end
-  end
-
   def repo(repo_name)
     client.repository(repo_name)
   end
 
-  def add_comment(options)
+  def add_pull_request_comment(options)
     client.create_pull_request_comment(
       options[:commit].repo_name,
       options[:pull_request_number],
@@ -42,24 +36,39 @@ class GithubApi
   def create_hook(full_repo_name, callback_endpoint)
     hook = client.create_hook(
       full_repo_name,
-      'web',
+      "web",
       { url: callback_endpoint },
-      { events: ['pull_request'], active: true }
+      { events: ["pull_request"], active: true }
     )
 
-    yield hook if block_given?
+    if block_given?
+      yield hook
+    else
+      hook
+    end
   rescue Octokit::UnprocessableEntity => error
-    unless error.message.include? 'Hook already exists'
+    if error.message.include? "Hook already exists"
+      true
+    else
       raise
     end
   end
 
   def remove_hook(full_github_name, hook_id)
-    client.remove_hook(full_github_name, hook_id)
+    response = client.remove_hook(full_github_name, hook_id)
+
+    if block_given?
+      yield
+    else
+      response
+    end
   end
 
   def pull_request_comments(full_repo_name, pull_request_number)
-    client.pull_request_comments(full_repo_name, pull_request_number)
+    repo_path = Octokit::Repository.path full_repo_name
+
+    # client.pull_request_comments does not do auto-pagination.
+    client.paginate "#{repo_path}/pulls/#{pull_request_number}/comments"
   end
 
   def pull_request_files(full_repo_name, number)
@@ -70,102 +79,83 @@ class GithubApi
     client.contents(full_repo_name, path: filename, ref: sha)
   end
 
+  def accept_pending_invitations
+    pending_memberships = client.organization_memberships(state: "pending")
+    pending_memberships.each do |pending_membership|
+      client.update_organization_membership(
+        pending_membership["organization"]["login"],
+        state: "active"
+      )
+    end
+  end
+
+  def create_pending_status(full_repo_name, sha, description)
+    create_status(
+      repo: full_repo_name,
+      sha: sha,
+      state: "pending",
+      description: description
+    )
+  end
+
+  def create_success_status(full_repo_name, sha, description)
+    create_status(
+      repo: full_repo_name,
+      sha: sha,
+      state: "success",
+      description: description
+    )
+  end
+
+  def add_collaborator(repo_name, username)
+    client.add_collaborator(repo_name, username)
+  end
+
   def user_teams
     client.user_teams
   end
 
-  def email_address
-    primary_email = client.emails.detect { |email| email['primary'] }
-    primary_email['email']
+  def repo_teams(repo_name)
+    client.repository_teams(repo_name)
+  end
+
+  def org_teams(org_name)
+    client.org_teams(org_name)
+  end
+
+  def create_team(team_name:, org_name:, repo_name:)
+    team_options = {
+      name: team_name,
+      repo_names: [repo_name],
+      permission: "push"
+    }
+    client.create_team(org_name, team_options)
+  end
+
+  def add_repo_to_team(team_id, repo_name)
+    client.add_team_repository(team_id, repo_name)
+  end
+
+  def add_user_to_team(username, team_id)
+    client.add_team_membership(team_id, username)
+  end
+
+  def update_team(team_id, options)
+    client.update_team(team_id, options)
   end
 
   private
 
-  def add_user_to_org(username, repo)
-    repo_teams = client.repository_teams(repo.full_name)
-    admin_team = admin_access_team(repo_teams)
-
-    if admin_team
-      add_user_to_team(username, admin_team.id)
-    else
-      add_user_and_repo_to_services_team(username, repo)
-    end
-  end
-
-  def admin_access_team(repo_teams)
-    token_bearer = GithubUser.new(self)
-
-    repo_teams.detect do |repo_team|
-      token_bearer.has_admin_access_through_team?(repo_team.id)
-    end
-  end
-
-  def add_user_and_repo_to_services_team(username, repo)
-    team = find_team(SERVICES_TEAM_NAME, repo)
-
-    if team
-      client.add_team_repository(team.id, repo.full_name)
-    else
-      team = create_team(SERVICES_TEAM_NAME, repo)
-    end
-
-    add_user_to_team(username, team.id)
-  end
-
-  def add_user_to_team(username, team_id)
-    client.add_team_member(team_id, username)
-  end
-
-  def find_team(name, repo)
-    client.org_teams(repo.organization.login).detect do |team|
-      team.name == name
-    end
-  end
-
-  def create_team(name, repo)
-    team_options = {
-      name: name,
-      repo_names: [repo.full_name],
-      permission: "pull"
-    }
-    client.create_team(repo.organization.login, team_options)
-  rescue Octokit::UnprocessableEntity => e
-    if team_exists_exception?(e)
-      find_team(name, repo)
-    else
-      raise
-    end
-  end
-
   def user_repos
-    repos = []
-    page = 1
-
-    loop do
-      results = client.repos(nil, page: page)
-      repos.concat(authorized_repos(results))
-      break unless results.any?
-      page += 1
-    end
-
-    repos
+    authorized_repos(client.repos)
   end
 
   def org_repos
-    repos = []
-
-    orgs.each do |org|
-      page = 1
-
-      loop do
-        results = client.org_repos(org[:login], page: page)
-        repos.concat(authorized_repos(results))
-        break unless results.any?
-        page += 1
-      end
+    repos = orgs.flat_map do |org|
+      client.org_repos(org[:login])
     end
 
-    repos
+    authorized_repos(repos)
   end
 
   def orgs
@@ -173,12 +163,22 @@ class GithubApi
   end
 
   def authorized_repos(repos)
-    repos.select {|repo| repo.permissions.admin }
+    repos.select { |repo| repo.permissions.admin }
   end
 
-  def team_exists_exception?(exception)
-    exception.errors.any? do |error|
-      error[:field] == "name" && error[:code] == "already_exists"
-    end
+  def with_preview_client(&block)
+    client.with_options(accept: PREVIEW_MEDIA_TYPE, &block)
+  end
+
+  def create_status(repo:, sha:, state:, description:)
+    client.create_status(
+      repo,
+      sha,
+      state,
+      context: "hound",
+      description: description
+    )
+  rescue Octokit::NotFound
+    # noop
   end
 end
