@@ -3,18 +3,20 @@ require "spec_helper"
 describe ActivationsController, "#create" do
   context "when activation succeeds" do
     it "returns successful response" do
+      token = "sometoken"
       membership = create(:membership)
       repo = membership.repo
       activator = double(:repo_activator, activate: true)
       allow(RepoActivator).to receive(:new).and_return(activator)
-      stub_sign_in(membership.user)
+      stub_sign_in(membership.user, token)
 
       post :create, repo_id: repo.id, format: :json
 
       expect(response.code).to eq "201"
       expect(response.body).to eq RepoSerializer.new(repo).to_json
-      expect(activator).to have_received(:activate).
-        with(repo, AuthenticationHelper::GITHUB_TOKEN)
+      expect(activator).to have_received(:activate)
+      expect(RepoActivator).to have_received(:new).
+        with(repo: repo, github_token: token)
       expect(analytics).to have_tracked("Activated Public Repo").
         for_user(membership.user).
         with(properties: { name: repo.full_github_name })
@@ -23,17 +25,19 @@ describe ActivationsController, "#create" do
 
   context "when activation fails" do
     it "returns error response" do
+      token = "sometoken"
       membership = create(:membership)
       repo = membership.repo
       activator = double(:repo_activator, activate: false).as_null_object
       allow(RepoActivator).to receive(:new).and_return(activator)
-      stub_sign_in(membership.user)
+      stub_sign_in(membership.user, token)
 
       post :create, repo_id: repo.id, format: :json
 
       expect(response.code).to eq "502"
-      expect(activator).to have_received(:activate).
-        with(repo, AuthenticationHelper::GITHUB_TOKEN)
+      expect(activator).to have_received(:activate)
+      expect(RepoActivator).to have_received(:new).
+        with(repo: repo, github_token: token)
     end
   end
 
@@ -47,7 +51,7 @@ describe ActivationsController, "#create" do
       stub_sign_in(user)
 
       expect { post :create, repo_id: repo.id, format: :json }.to raise_error(
-        ActivationsController::CannotActivatePrivateRepo
+        ActivationsController::CannotActivatePaidRepo
       )
       expect(activator).not_to have_received(:activate)
     end
