@@ -1,23 +1,23 @@
-require 'spec_helper'
+require "spec_helper"
 
-describe SessionsController, '#new' do
-  context 'when https is enabled' do
-    context 'and http is used' do
-      it 'redirects to https' do
+describe HomeController, "#index" do
+  context "when https is enabled" do
+    context "and http is used" do
+      it "redirects to https" do
         with_https_enabled do
-          get :new
+          get :index
 
-          expect(response).to redirect_to(sign_in_url(protocol: 'https'))
+          expect(response).to redirect_to(root_url(protocol: "https"))
         end
       end
     end
 
-    context 'and https is used' do
-      it 'does not redirect' do
+    context "and https is used" do
+      it "does not redirect" do
         with_https_enabled do
-          request.env['HTTPS'] = 'on'
+          request.env["HTTPS"] = "on"
 
-          get :new
+          get :index
 
           expect(response).not_to be_redirect
         end
@@ -25,10 +25,10 @@ describe SessionsController, '#new' do
     end
   end
 
-  context 'when https is disabled' do
-    context 'and http is used' do
-      it 'does not redirect' do
-        get :new
+  context "when https is disabled" do
+    context "and http is used" do
+      it "does not redirect" do
+        get :index
 
         expect(response).not_to be_redirect
       end
