@@ -1,49 +1,55 @@
 require "spec_helper"
 
 describe SubscriptionsController, "#create" do
-  it "subscribes the user to the repo" do
-    membership = create(:membership)
-    repo = membership.repo
-    activator = double(:repo_activator, activate: true)
-    allow(RepoActivator).to receive(:new).and_return(activator)
-    allow(RepoSubscriber).to receive(:subscribe).and_return(true)
-    stub_sign_in(membership.user)
+  context "when subscription succeeds" do
+    it "subscribes the user to the repo" do
+      token = "usergithubtoken"
+      membership = create(:membership)
+      repo = membership.repo
+      activator = double(:repo_activator, activate: true)
+      allow(RepoActivator).to receive(:new).and_return(activator)
+      allow(RepoSubscriber).to receive(:subscribe).and_return(true)
+      allow(JobQueue).to receive(:push)
+      stub_sign_in(membership.user, token)
 
-    post(
-      :create,
-      repo_id: repo.id,
-      card_token: "cardtoken",
-      email_address: "jimtom@example.com",
-      format: :json
-    )
+      post(
+        :create,
+        repo_id: repo.id,
+        card_token: "cardtoken",
+        email_address: "jimtom@example.com",
+        format: :json
+      )
 
-    expect(activator).to have_received(:activate).
-      with(repo, AuthenticationHelper::GITHUB_TOKEN)
-    expect(RepoSubscriber).to have_received(:subscribe).
-      with(repo, membership.user, "cardtoken")
-    expect(analytics).to have_tracked("Subscribed Private Repo").
-      for_user(membership.user).
-      with(properties: { name: repo.name, revenue: repo.plan_price })
-  end
+      expect(activator).to have_received(:activate)
+      expect(RepoActivator).to have_received(:new).
+        with(repo: repo, github_token: token)
+      expect(RepoSubscriber).to have_received(:subscribe).
+        with(repo, membership.user, "cardtoken")
+      expect(analytics).to have_tracked("Subscribed Private Repo").
+        for_user(membership.user).
+        with(properties: { name: repo.name, revenue: repo.plan_price })
+    end
 
-  it "updates the current user's email address" do
-    user = create(:user, email_address: nil)
-    repo = create(:repo)
-    user.repos << repo
-    activator = double(:repo_activator, activate: true)
-    allow(RepoActivator).to receive(:new).and_return(activator)
-    allow(RepoSubscriber).to receive(:subscribe).and_return(true)
-    stub_sign_in(user)
+    it "updates the current user's email address" do
+      user = create(:user, email_address: nil)
+      repo = create(:repo)
+      user.repos << repo
+      activator = double(:repo_activator, activate: true)
+      allow(RepoActivator).to receive(:new).and_return(activator)
+      allow(RepoSubscriber).to receive(:subscribe).and_return(true)
+      allow(JobQueue).to receive(:push)
+      stub_sign_in(user)
 
-    post(
-      :create,
-      repo_id: repo.id,
-      card_token: "cardtoken",
-      email_address: "jimtom@example.com",
-      format: :json
-    )
+      post(
+        :create,
+        repo_id: repo.id,
+        card_token: "cardtoken",
+        email_address: "jimtom@example.com",
+        format: :json
+      )
 
-    expect(user.reload.email_address).to eq "jimtom@example.com"
+      expect(user.reload.email_address).to eq "jimtom@example.com"
+    end
   end
 
   context "when subscription fails" do
@@ -65,6 +71,7 @@ end
 
 describe SubscriptionsController, "#destroy" do
   it "deletes subscription associated with subscribing user" do
+    token = "usertoken"
     current_user = create(:user)
     subscribed_user = create(:user)
     membership = create(:membership, user: current_user)
@@ -73,7 +80,7 @@ describe SubscriptionsController, "#destroy" do
     activator = double(:repo_activator, deactivate: true)
     allow(RepoActivator).to receive(:new).and_return(activator)
     allow(RepoSubscriber).to receive(:unsubscribe).and_return(true)
-    stub_sign_in(current_user)
+    stub_sign_in(current_user, token)
 
     delete(
       :destroy,
@@ -82,8 +89,9 @@ describe SubscriptionsController, "#destroy" do
       format: :json
     )
 
-    expect(activator).to have_received(:deactivate).
-      with(repo, AuthenticationHelper::GITHUB_TOKEN)
+    expect(activator).to have_received(:deactivate)
+    expect(RepoActivator).to have_received(:new).
+      with(repo: repo, github_token: token)
     expect(RepoSubscriber).to have_received(:unsubscribe).
       with(repo, subscribed_user)
     expect(analytics).to have_tracked("Unsubscribed Private Repo").
