@@ -14,7 +14,7 @@ feature 'Builds' do
     scenario 'a successful build with custom config' do
       create(:repo, github_id: repo_id, full_github_name: repo_name)
       stub_github_requests_with_no_violations
-      comment_request = stubbed_comment_request
+      comment_request = stub_simple_comment_request
 
       page.driver.post builds_path, payload: payload
 
@@ -26,7 +26,7 @@ feature 'Builds' do
     scenario 'a successful build with custom config' do
       create(:repo, github_id: repo_id, full_github_name: repo_name)
       stub_github_requests_with_no_violations
-      comment_request = stubbed_comment_request
+      comment_request = stub_simple_comment_request
 
       page.driver.post builds_path, payload
 
@@ -37,13 +37,10 @@ feature 'Builds' do
   scenario 'a failed build' do
     create(:repo, :active, github_id: repo_id, full_github_name: repo_name)
     stub_github_requests_with_violations
-    stub_commit_request(repo_name, pr_sha, ENV["HOUND_GITHUB_TOKEN"])
-    stub_pull_request_comments_request(
-      repo_name,
-      pr_number,
-      ENV['HOUND_GITHUB_TOKEN']
-    )
-    comment_request = stubbed_comment_request
+    stub_commit_request(repo_name, pr_sha)
+    stub_pull_request_comments_request(repo_name, pr_number)
+    comment_request = stub_simple_comment_request
+    stub_status_requests(repo_name, pr_sha)
 
     page.driver.post builds_path, payload: payload
 
@@ -51,20 +48,14 @@ feature 'Builds' do
   end
 
   def stub_github_requests_with_no_violations
-    stub_pull_request_files_request(
-      repo_name,
-      pr_number,
-      ENV['HOUND_GITHUB_TOKEN']
-    )
+    stub_pull_request_files_request(repo_name, pr_number)
     stub_contents_request(
-      ENV['HOUND_GITHUB_TOKEN'],
       repo_name: repo_name,
       sha: pr_sha,
       file: 'file1.rb',
       fixture: 'contents.json'
     )
     stub_contents_request(
-      ENV['HOUND_GITHUB_TOKEN'],
       repo_name: repo_name,
       sha: pr_sha,
       file: '.hound.yml',
@@ -73,27 +64,21 @@ feature 'Builds' do
   end
 
   def stub_github_requests_with_violations
-    stub_pull_request_files_request(
-      repo_name,
-      pr_number,
-      ENV['HOUND_GITHUB_TOKEN']
-    )
+    stub_pull_request_files_request(repo_name, pr_number)
     stub_contents_request(
-      ENV['HOUND_GITHUB_TOKEN'],
       repo_name: repo_name,
       sha: pr_sha,
       file: '.hound.yml',
       fixture: 'config_contents.json'
     )
     stub_contents_request(
-      ENV["HOUND_GITHUB_TOKEN"],
       repo_name: repo_name,
       sha: pr_sha,
       fixture: 'contents_with_violations.json'
     )
   end
 
-  def stubbed_comment_request
+  def stub_simple_comment_request
     stub_request(
       :post,
       "https://api.github.com/repos/#{repo_name}/pulls/#{pr_number}/comments"
