@@ -2,8 +2,9 @@ require 'spec_helper'
 
 feature 'User authentication' do
   scenario "existing user signs in" do
+    token = "usergithubtoken"
     user = create(:user)
-    stub_repo_requests(AuthenticationHelper::GITHUB_TOKEN)
+    stub_repo_requests(token)
 
     sign_in_as(user)
 
@@ -12,20 +13,22 @@ feature 'User authentication' do
   end
 
   scenario "new user signs in" do
+    token = "usergithubtoken"
     github_username = "croaky"
     user = build(:user, github_username: github_username)
-    stub_repo_requests(AuthenticationHelper::GITHUB_TOKEN)
+    stub_repo_requests(token)
 
-    sign_in_as(user)
+    sign_in_as(user, token)
 
     expect(page).to have_content(github_username)
   end
 
   scenario 'user signs out' do
+    token = "usergithubtoken"
     user = create(:user)
-    stub_repo_requests(AuthenticationHelper::GITHUB_TOKEN)
-    sign_in_as(user)
+    stub_repo_requests(token)
 
+    sign_in_as(user, token)
     find('a[href="/sign_out"]').click
 
     expect(page).not_to have_content user.github_username
