@@ -7,10 +7,9 @@ describe RepoInformationJob do
 
   it 'collects repo privacy and organization from GitHub' do
     repo = create(:repo, private: false, in_organization: false)
-    github_token = 'token'
-    stub_repo_with_org_request(repo.full_github_name, github_token)
+    stub_repo_with_org_request(repo.full_github_name)
 
-    RepoInformationJob.perform(repo.id, github_token)
+    RepoInformationJob.perform(repo.id)
 
     repo.reload
     expect(repo).to be_private
@@ -18,14 +17,12 @@ describe RepoInformationJob do
   end
 
   it 'retries when Resque::TermException is raised' do
-    user_id = 'userid'
-    github_token = 'token'
-    allow(User).to receive(:find).and_raise(Resque::TermException.new(1))
+    repo = create(:repo)
+    allow(Repo).to receive(:find).and_raise(Resque::TermException.new(1))
     allow(Resque).to receive(:enqueue)
 
-    EmailAddressJob.perform(user_id, github_token)
+    RepoInformationJob.perform(repo.id)
 
-    expect(Resque).to have_received(:enqueue).
-      with(EmailAddressJob, user_id, github_token)
+    expect(Resque).to have_received(:enqueue).with(RepoInformationJob, repo.id)
   end
 end
