@@ -1,11 +1,13 @@
-require "spec_helper"
+require "fast_spec_helper"
+require "app/models/pull_request"
+require "app/models/commit"
+require "lib/github_api"
 
 describe PullRequest do
   describe "#opened?" do
     context "when payload action is opened" do
       it "returns true" do
-        payload = double(:payload, action: "opened")
-        pull_request = PullRequest.new(payload, "token")
+        pull_request = PullRequest.new(payload_stub(action: "opened"))
 
         expect(pull_request).to be_opened
       end
@@ -13,8 +15,8 @@ describe PullRequest do
 
     context "when payload action is not opened" do
       it "returns false" do
-        payload = double(:payload, action: "notopened")
-        pull_request = PullRequest.new(payload, "token")
+        payload = payload_stub(action: "notopened")
+        pull_request = PullRequest.new(payload)
 
         expect(pull_request).not_to be_opened
       end
@@ -24,8 +26,8 @@ describe PullRequest do
   describe "#synchronize?" do
     context "when payload action is synchronize" do
       it "returns true" do
-        payload = double(:payload, action: "synchronize")
-        pull_request = PullRequest.new(payload, "token")
+        payload = payload_stub(action: "synchronize")
+        pull_request = PullRequest.new(payload)
 
         expect(pull_request).to be_synchronize
       end
@@ -33,8 +35,8 @@ describe PullRequest do
 
     context "when payload action is not synchronize" do
       it "returns false" do
-        payload = double(:payload, action: "notsynchronize")
-        pull_request = PullRequest.new(payload, "token")
+        payload = payload_stub(action: "notsynchronize")
+        pull_request = PullRequest.new(payload)
 
         expect(pull_request).not_to be_synchronize
       end
@@ -43,18 +45,10 @@ describe PullRequest do
 
   describe "#comments" do
     it "returns comments on pull request" do
-      payload = double(
-        :payload,
-        full_repo_name: "org/repo",
-        number: 4,
-        head_sha: "abc123"
-      )
-      patch_position = 7
       filename = "spec/models/style_guide_spec.rb"
-      comment = double(:comment, position: patch_position, path: filename)
+      comment = double(:comment, position: 7, path: filename)
       github = double(:github, pull_request_comments: [comment])
-      allow(GithubApi).to receive(:new).and_return(github)
-      pull_request = PullRequest.new(payload, "githubtoken")
+      pull_request = pull_request_stub(github)
 
       comments = pull_request.comments
 
@@ -63,46 +57,47 @@ describe PullRequest do
     end
   end
 
-  describe "#add_comment" do
+  describe "#comment_on_violation" do
     it "posts a comment to GitHub for the Hound user" do
-      payload = double(
-        :payload,
-        full_repo_name: "org/repo",
-        number: "123",
-        head_sha: "1234abcd"
-      )
-      commit = double(:commit, repo_name: payload.full_repo_name)
-      github = double(:github_client, add_comment: nil)
-      allow(GithubApi).to receive(:new).and_return(github)
+      payload = payload_stub
+      github = double(:github_client, add_pull_request_comment: nil)
+      pull_request = pull_request_stub(github, payload)
+      violation = violation_stub
+      commit = double("Commit")
       allow(Commit).to receive(:new).and_return(commit)
-      violation = double(
-        :violation,
-        messages: ["A comment"],
-        filename: "test.rb",
-        patch_position: 123,
-      )
-      pull_request = PullRequest.new(payload, "gh-token")
 
-      pull_request.add_comment(violation)
+      pull_request.comment_on_violation(violation)
 
-      expect(github).to have_received(:add_comment).with(
-        pull_request_number: payload.number,
+      expect(github).to have_received(:add_pull_request_comment).with(
+        pull_request_number: payload.pull_request_number,
         commit: commit,
-        comment: "A comment",
-        filename: "test.rb",
-        patch_position: 123
+        comment: violation.messages.first,
+        filename: violation.filename,
+        patch_position: violation.patch_position,
       )
     end
   end
 
-  def pull_request(api)
-    payload = double(
-      :payload,
-      number: 1,
+  def violation_stub(options = {})
+    defaults =  {
+      messages: ["A comment"],
+      filename: "test.rb",
+      patch_position: 123,
+    }
+    double("Violation", defaults.merge(options))
+  end
+
+  def payload_stub(options = {})
+    defaults = {
       full_repo_name: "org/repo",
-      head_sha: "abc123"
-    )
+      head_sha: "1234abcd",
+      pull_request_number: 5,
+    }
+    double("Payload", defaults.merge(options))
+  end
+
+  def pull_request_stub(api, payload = payload_stub)
     allow(GithubApi).to receive(:new).and_return(api)
-    PullRequest.new(payload, "gh-token")
+    PullRequest.new(payload)
   end
 end
