@@ -1,16 +1,4 @@
-require "active_support/core_ext"
-require "coffeelint"
-require "jshintrb"
-require "rubocop"
-
-require "fast_spec_helper"
-require "app/models/line"
-require "app/models/unchanged_line"
-require "app/models/repo_config"
-require "app/models/style_checker"
-require "app/models/violation"
-require "app/models/violations"
-Dir.glob("app/models/style_guide/*.rb", &method(:require))
+require "spec_helper"
 
 describe StyleChecker, "#violations" do
   it "returns a collection of computed violations" do
@@ -65,6 +53,18 @@ describe StyleChecker, "#violations" do
   end
 
   context "for a CoffeeScript file" do
+    it "is processed with a coffee.js extension" do
+      file = stub_commit_file("test.coffee.js", "foo ->")
+      pull_request = stub_pull_request(pull_request_files: [file])
+      style_checker = StyleChecker.new(pull_request)
+      allow(RepoConfig).to receive(:new).and_return(stub_repo_config)
+
+      violations = style_checker.violations
+      messages = violations.flat_map(&:messages)
+
+      expect(messages).to eq ["Empty function"]
+    end
+
     context "with violations" do
       context "with CoffeeScript enabled" do
         it "returns violations" do
@@ -73,7 +73,7 @@ describe StyleChecker, "#violations" do
               enabled: true
           YAML
           head_commit = double("Commit", file_content: config)
-          file = stub_commit_file("test.coffee", "alert 'Hello World'")
+          file = stub_commit_file("test.coffee", "foo: ->")
           pull_request = stub_pull_request(
             head_commit: head_commit,
             pull_request_files: [file],
@@ -82,7 +82,7 @@ describe StyleChecker, "#violations" do
           violations = StyleChecker.new(pull_request).violations
           messages = violations.flat_map(&:messages)
 
-          expect(messages).to eq ["Implicit parens are forbidden"]
+          expect(messages).to eq ["Empty function"]
         end
       end
 
@@ -191,6 +191,94 @@ describe StyleChecker, "#violations" do
         end
       end
     end
+
+    context "an excluded file" do
+      it "returns no violations" do
+        config = <<-YAML.strip_heredoc
+          java_script:
+            enabled: true
+            ignore_file: '.jshintignore'
+        YAML
+
+        head_commit = stub_head_commit(
+          ".hound.yml" => config,
+          ".jshintignore" => "test.js"
+        )
+
+        file = stub_commit_file("test.js", "var test = 'test'")
+        pull_request = stub_pull_request(
+          head_commit: head_commit,
+          pull_request_files: [file]
+        )
+
+        violations = StyleChecker.new(pull_request).violations
+
+        expect(violations).to be_empty
+      end
+    end
+  end
+
+  context "for a SCSS file" do
+    context "with violations" do
+      context "with SCSS enabled" do
+        it "returns violations" do
+          head_commit = stub_head_commit(".hound.yml" => scss_enabled_config)
+          file = stub_commit_file(
+            "test.scss",
+            ".table p.inner table td { background: red; }"
+          )
+          pull_request = stub_pull_request(
+            head_commit: head_commit,
+            pull_request_files: [file],
+          )
+
+          violations = StyleChecker.new(pull_request).violations
+          messages = violations.flat_map(&:messages)
+
+          expect(messages).to include(
+            "Selector should have depth of applicability no greater than 2, but was 4"
+          )
+        end
+      end
+
+      context "with SCSS disabled" do
+        it "returns no violations" do
+          head_commit = stub_head_commit(".hound.yml" => scss_disabled_config)
+          file = stub_commit_file(
+            "test.scss",
+            ".table p.inner table td { background: red; }"
+          )
+          pull_request = stub_pull_request(
+            head_commit: head_commit,
+            pull_request_files: [file],
+          )
+
+          violations = StyleChecker.new(pull_request).violations
+
+          expect(violations).to be_empty
+        end
+      end
+    end
+
+    context "without violations" do
+      context "with SCSS enabled" do
+        it "returns no violations" do
+          head_commit = stub_head_commit(".hound.yml" => scss_enabled_config)
+          file = stub_commit_file("test.scss", "table td { color: green; }")
+          pull_request = stub_pull_request(
+            head_commit: head_commit,
+            pull_request_files: [file],
+          )
+
+          violations = StyleChecker.new(pull_request).violations
+          messages = violations.flat_map(&:messages)
+
+          expect(messages).not_to include(
+            "Selector should have depth of applicability no greater than 3"
+          )
+        end
+      end
+    end
   end
 
   context "with unsupported file type" do
@@ -204,6 +292,18 @@ describe StyleChecker, "#violations" do
     end
   end
 
+  context "a removed file" do
+    it "does not return a violation for the file" do
+      file = stub_commit_file("ruby.rb", "puts 123    ", removed: true)
+      pull_request = stub_pull_request(pull_request_files: [file])
+
+      violations = StyleChecker.new(pull_request).violations
+      messages = violations.flat_map(&:messages)
+
+      expect(messages).to eq []
+    end
+  end
+
   private
 
   def stub_pull_request(options = {})
@@ -212,20 +312,55 @@ describe StyleChecker, "#violations" do
       file_content: "",
       head_commit: head_commit,
       pull_request_files: [],
+      repository_owner: "some_org"
     }
 
     double("PullRequest", defaults.merge(options))
   end
 
-  def stub_commit_file(filename, contents, line = nil)
+  def stub_commit_file(filename, contents, line = nil, removed: false)
     line ||= Line.new(content: "foo", number: 1, patch_position: 2)
     formatted_contents = "#{contents}\n"
     double(
       filename.split(".").first,
       filename: filename,
       content: formatted_contents,
-      removed?: false,
+      removed?: removed,
       line_at: line,
     )
+  end
+
+  def stub_head_commit(options)
+    head_commit = double("Commit", file_content: nil)
+
+    options.each do |filename, file_contents|
+      allow(head_commit).to receive(:file_content).
+        with(filename).and_return(file_contents)
+    end
+
+    head_commit
+  end
+
+  def stub_repo_config
+    double(
+      "RepoConfig",
+      for: {},
+      enabled_for?: true,
+      ignored_javascript_files: []
+    )
+  end
+
+  def scss_enabled_config
+    <<-YAML.strip_heredoc
+      scss:
+        enabled: true
+    YAML
+  end
+
+  def scss_disabled_config
+    <<-YAML.strip_heredoc
+      scss:
+        enabled: false
+    YAML
   end
 end
