@@ -3,61 +3,53 @@ require 'spec_helper'
 describe RepoActivator do
   describe '#activate' do
     context 'when repo activation succeeds' do
-      it 'activates repo' do
-        github_token = 'githubtoken'
+      it "marks repo as active" do
         repo = create(:repo)
         stub_github_api
-        activator = RepoActivator.new
+        activator = build_activator(repo: repo)
 
-        expect(activator.activate(repo, github_token)).to be_truthy
-        expect(GithubApi).to have_received(:new).with(github_token)
+        result = activator.activate
+
+        expect(result).to be_truthy
         expect(repo.reload).to be_active
       end
 
-      it 'makes Hound a collaborator' do
-        repo = create(:repo)
+      it "adds Hound to the repo" do
+        token = "some_token"
+        repo = build(:repo, name: "foo/bar")
         github = stub_github_api
-        activator = RepoActivator.new
+        activator = build_activator(repo: repo, token: token)
 
-        activator.activate(repo, 'githubtoken')
+        activator.activate
 
-        expect(github).to have_received(:add_user_to_repo)
+        expect(GithubApi).to have_received(:new).with(token)
+        expect(AddHoundToRepo).to have_received(:run).with(repo.name, github)
       end
 
-      it 'returns true if the repo activates successfully' do
-        repo = create(:repo)
-        stub_github_api
-        activator = RepoActivator.new
+      context "when https is enabled" do
+        it "creates GitHub hook using secure build URL" do
+          github = stub_github_api
+          repo = build(:repo, name: "foo/bar")
+          activator = build_activator(repo: repo)
 
-        response = activator.activate(repo, 'githubtoken')
-
-        expect(response).to be_truthy
-      end
-
-      context 'when https is enabled' do
-        it 'creates GitHub hook using secure build URL' do
           with_https_enabled do
-            repo = create(:repo)
-            github = stub_github_api
-            activator = RepoActivator.new
-
-            activator.activate(repo, 'githubtoken')
-
-            expect(github).to have_received(:create_hook).with(
-              repo.full_github_name,
-              URI.join("https://#{ENV['HOST']}", 'builds').to_s
-            )
+            activator.activate
           end
+
+          expect(github).to have_received(:create_hook).with(
+            repo.full_github_name,
+            URI.join("https://#{ENV['HOST']}", "builds").to_s
+          )
         end
       end
 
       context 'when https is disabled' do
         it 'creates GitHub hook using insecure build URL' do
-          repo = create(:repo)
           github = stub_github_api
-          activator = RepoActivator.new
+          repo = build(:repo)
+          activator = build_activator(repo: repo)
 
-          activator.activate(repo, 'githubtoken')
+          activator.activate
 
           expect(github).to have_received(:create_hook).with(
             repo.full_github_name,
@@ -67,118 +59,115 @@ describe RepoActivator do
       end
     end
 
-    context 'when repo activation fails' do
-      it 'returns false if API request raises' do
-        github_token = nil
-        repo = double('repo')
-        expect(GithubApi).to receive(:new).and_raise(Octokit::Error.new)
-        activator = RepoActivator.new
+    context "when adding hound to rope results in an error" do
+      it "returns false" do
+        activator = build_activator
+        allow(AddHoundToRepo).to receive(:run).and_raise(Octokit::Error.new)
 
-        response = activator.activate(repo, github_token)
+        result = activator.activate
 
-        expect(response).to be_falsy
+        expect(result).to be_falsy
       end
 
-      it 'only swallows Octokit errors' do
-        github_token = 'githubtoken'
-        repo = double('repo')
-        expect(GithubApi).to receive(:new).and_raise(Exception.new)
-        activator = RepoActivator.new
+      it "reports raised exception to Sentry" do
+        activator = build_activator
+        error = Octokit::Error.new
+        allow(AddHoundToRepo).to receive(:run).and_raise(error)
+        allow(Raven).to receive(:capture_exception)
 
-        expect { activator.activate(repo, github_token) }.to raise_error(Exception)
+        activator.activate
+
+        expect(Raven).to have_received(:capture_exception).with(error)
       end
 
-      context 'when Hound cannot be added to repo' do
-        it 'returns false' do
-          repo = double(:repo, full_github_name: 'test/repo')
-          github = double(:github, add_user_to_repo: false)
-          allow(GithubApi).to receive(:new).and_return(github)
-          activator = RepoActivator.new
+      it "only swallows Octokit errors" do
+        activator = build_activator
+        allow(AddHoundToRepo).to receive(:run).and_raise(StandardError.new)
 
-          expect(activator.activate(repo, github)).to be_falsy
-        end
+        expect { activator.activate }.to raise_error(StandardError)
       end
     end
 
     context 'hook already exists' do
       it 'does not raise' do
-        token = 'token'
-        repo = create(:repo)
-        github = double(:github, create_hook: nil, add_user_to_repo: true)
+        activator = build_activator
+        github = double("GithubApi", create_hook: nil)
         allow(GithubApi).to receive(:new).and_return(github)
-        activator = RepoActivator.new
 
-        expect { activator.activate(repo, token) }.not_to raise_error
-
-        expect(GithubApi).to have_received(:new).with(token)
+        expect { activator.activate }.not_to raise_error
       end
     end
   end
 
-  describe '#deactivate' do
-    context 'when repo activation succeeds' do
-      it 'deactivates repo' do
-        stub_github_api
-        github_token = 'githubtoken'
+  describe "#deactivate" do
+    context "when repo deactivation succeeds" do
+      it "marks repo as deactivated" do
         repo = create(:repo)
         create(:membership, repo: repo)
-        activator = RepoActivator.new
+        activator = build_activator(repo: repo)
+        stub_github_api
 
-        activator.deactivate(repo, github_token)
+        activator.deactivate
 
-        expect(GithubApi).to have_received(:new).with(github_token)
-        expect(repo.active?).to be_falsy
+        expect(repo.reload).not_to be_active
       end
 
       it 'removes GitHub hook' do
-        github_api = stub_github_api
         repo = create(:repo)
         create(:membership, repo: repo)
-        activator = RepoActivator.new
+        activator = build_activator(repo: repo)
+        github_api = stub_github_api
 
-        activator.deactivate(repo, 'githubtoken')
+        activator.deactivate
 
         expect(github_api).to have_received(:remove_hook)
         expect(repo.hook_id).to be_nil
       end
 
-      it 'returns true if the repo activates successfully' do
+      it "returns true" do
         stub_github_api
+        token = "githubtoken"
+        allow(JobQueue).to receive(:push)
         membership = create(:membership)
-        activator = RepoActivator.new
+        repo = membership.repo
+        activator = RepoActivator.new(github_token: token, repo: repo)
 
-        response = activator.deactivate(membership.repo, "githubtoken")
+        result = activator.deactivate
 
-        expect(response).to be_truthy
+        expect(result).to be_truthy
       end
     end
 
-    context 'when repo activation succeeds' do
-      it 'returns false if the repo does not activate successfully' do
-        repo = double('repo')
-        github_token = nil
-        expect(GithubApi).to receive(:new).and_raise(Octokit::Error.new)
-        activator = RepoActivator.new
+    context "when repo deactivation fails" do
+      it "returns false" do
+        activator = build_activator
+        allow(GithubApi).to receive(:new).and_raise(Octokit::Error.new)
 
-        response = activator.deactivate(repo, github_token)
+        result = activator.deactivate
 
-        expect(response).to be_falsy
+        expect(result).to be_falsy
       end
 
       it 'only swallows Octokit errors' do
-        repo = double('repo')
-        github_token = nil
-        expect(GithubApi).to receive(:new).and_raise(Exception.new)
-        activator = RepoActivator.new
+        error = StandardError.new("this should bubble through")
+        activator = build_activator
+        expect(GithubApi).to receive(:new).and_raise(error)
 
-        expect { activator.deactivate(repo, github_token) }.to raise_error(Exception)
+        expect { activator.deactivate }.to raise_error(error)
       end
     end
   end
 
+  def build_activator(token: "githubtoken", repo: build(:repo))
+    allow(JobQueue).to receive(:push).and_return(true)
+    allow(AddHoundToRepo).to receive(:run).and_return(true)
+
+    RepoActivator.new(github_token: token, repo: repo)
+  end
+
   def stub_github_api
     hook = double(:hook, id: 1)
-    api = double(:github_api, add_user_to_repo: true, remove_hook: true)
+    api = double(:github_api, remove_hook: true)
     allow(api).to receive(:create_hook).and_yield(hook)
     allow(GithubApi).to receive(:new).and_return(api)
     api
